requirements.txt
langchain~=0.3.14
openai
requests~=2.32.3
langchain-openai~=0.3.0
python-dotenv~=1.0.1
langchain-core~=0.3.29
langgraph~=0.2.62